refactor : created a common implementation of the RequestBuilder
